+++ components/form/admin/submission.values.tsx
@@ -7,6 +7,7 @@ import {
   SubmissionFragment,
   SubmissionFragmentField,
 } from '../../../graphql/fragment/submission.fragment'
+import { fieldTypes } from '../types'
 
 interface Props {
   form: FormPagerFragment
@@ -30,21 +31,8 @@ export const SubmissionValues: React.FC<Props> = (props) => {
     {
       title: t('submission:value'),
       render(_, row) {
-        console.log('row.value', row.value)
-
         try {
-          const data = JSON.parse(row.value) as { value: string }
-
-          if (Array.isArray(data.value)) {
-            return (
-              <ul>
-                {data.value.map(r => (
-                  <li key={r}>{JSON.stringify(r)}</li>
-                ))}
-              </ul>
-            )
-          }
-          return data.value
+          return fieldTypes[row.type]?.displayValue(row.value)
         } catch (e) {
           return row.value
         }