+++ components/form/layouts/card/index.tsx
@@ -139,7 +139,14 @@ export const CardLayout: React.FC<LayoutProps> = (props) => {
                   return null
                 }
 
-                return <Field key={field.id} field={field} design={design} focus={i === 0} />
+                return (
+                  <Field
+                    key={field.id}
+                    field={field}
+                    design={design}
+                    focus={false}
+                  />
+                )
               })}
               <div
                 style={{