+++ pages/admin/forms/create.tsx
@@ -82,7 +82,6 @@ const Create: NextPage = () => {
           noStyle
           name={[
             'form',
-            'design',
             'layout',
           ]}
           initialValue={'card'}