+++ CHANGELOG.md
@@ -16,11 +16,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - update translations (https://github.com/ohmyform/ui/pull/70)
 - show warning icon in form list if not public
 - default form layout is now "card"
+- creating of new fields combined in new field types
 
 ### Fixed
 
 - locale scripts were missing dependency
 - edit user shows now email in title
+- focus is now passed also do slide layout fields
+- empty fields are no longer submitted
 
 ### Security